app/requirements.txt
paho-mqtt==2.1.0
pymodbus==3.7.4
aiofiles==24.1.0
attrs==24.2.0
PyYAML==6.0.1
